+++ tests/wps/test_utils.py
@@ -57,7 +57,7 @@ def test_get_wps_client_headers_preserved():
     assert wps.url == test_wps_url
 
 
-def test_get_wps_output_context():
+def test_get_wps_output_context_validation():
     bad_cases = [
         "test/////test",
         "test/test//",
@@ -89,13 +89,72 @@ def test_get_wps_output_context():
 
     for header in header_names:
         for case in bad_cases:
-            with pytest.raises(HTTPUnprocessableEntity):
+            # validate against provided header
+            try:
                 req = DummyRequest(headers={header: case})
                 get_wps_output_context(req)
+            except HTTPUnprocessableEntity:
+                pass
+            else:
+                pytest.fail("Exception not raised when expected: (header={}, case={})".format(header, case))
+            # validate same conditions fulfilled by default context if header omitted
+            try:
+                settings = {"weaver.wps_output_context": case}
+                with mock.patch("weaver.wps.utils.get_settings", return_value=settings):
+                    req = DummyRequest(headers={})
+                    get_wps_output_context(req)
+            except HTTPUnprocessableEntity:
+                pass
+            else:
+                pytest.fail("Exception not raised when expected: (<setting>, case={})".format(case))
+
         for case, result in good_cases:
+            # validate against provided header
             try:
                 req = DummyRequest(headers={header: case})
                 ctx = get_wps_output_context(req)
                 assert ctx == result
             except Exception as exc:
-                pytest.fail("Exception raised when none is expected: {!s}: ${!s}".format(exc.__class__.__name__, exc))
+                pytest.fail("Exception raised when none is expected: (header={}, case={})\n"
+                            "Exception: {!s}: ${!s}".format(header, case, exc.__class__.__name__, exc))
+
+            # validate same conditions fulfilled by default context if header omitted
+            settings = {"weaver.wps_output_context": case}
+            with mock.patch("weaver.wps.utils.get_settings", return_value=settings):
+                try:
+                    req = DummyRequest(headers={})
+                    ctx = get_wps_output_context(req)
+                    assert ctx == result
+                except Exception as exc:
+                    pytest.fail("Exception raised when none is expected: (<setting>, case={})\n"
+                                "Exception: {!s}: ${!s}".format(case, exc.__class__.__name__, exc))
+
+
+def test_get_wps_output_context_resolution():
+    test_cases = [
+        # header provided uses it regardless of setting
+        ("somewhere", None, "somewhere"),
+        ("somewhere", "", "somewhere"),
+        ("somewhere", "test", "somewhere"),
+        # header empty or omitted defaults to none used when no/empty default context
+        ("", None, None),
+        ("", "", None),
+        (None, None, None),
+        (None, "", None),
+        # header empty or omitted defaults to defined default context setting
+        ("", "test", "test"),
+        (None, "test", "test"),
+    ]
+
+    for i, (test_header, test_setting, expect_result) in enumerate(test_cases):
+        try:
+            headers = {"X-WPS-Output-Context": test_header}
+            settings = {"weaver.wps_output_context": test_setting}
+            req = DummyRequest(headers=headers)
+            with mock.patch("weaver.wps.utils.get_settings", return_value=settings):
+                res = get_wps_output_context(req)
+            assert res == expect_result
+        except Exception as exc:
+            pytest.fail(
+                "Exception raised when none is expected [{}]: {!s}: ${!s}".format(i, exc.__class__.__name__, exc)
+            )